standardize function source code location... move to handlers subdir (#53)
