+++ services/app-api/handlers/create.js
@@ -1,6 +1,6 @@
 import * as uuid from "uuid";
-import handler from "./libs/handler-lib";
-import dynamoDb from "./libs/dynamodb-lib";
+import handler from "./../libs/handler-lib";
+import dynamoDb from "./../libs/dynamodb-lib";
 
 export const main = handler(async (event, context) => {
   // If this invokation is a prewarm, do nothing and return.
