+++ services/app-api/handlers/delete.js
@@ -1,5 +1,5 @@
-import handler from "./libs/handler-lib";
-import dynamoDb from "./libs/dynamodb-lib";
+import handler from "./../libs/handler-lib";
+import dynamoDb from "./../libs/dynamodb-lib";
 
 export const main = handler(async (event, context) => {
   // If this invokation is a prewarm, do nothing and return.
@@ -10,7 +10,7 @@ export const main = handler(async (event, context) => {
 
   const params = {
     TableName: process.env.tableName,
-    // 'Key' defines the partition key and sort key of the item to be retrieved
+    // 'Key' defines the partition key and sort key of the item to be removed
     // - 'userId': Identity Pool identity id of the authenticated user
     // - 'amendmentId': path parameter
     Key: {
@@ -19,11 +19,7 @@ export const main = handler(async (event, context) => {
     }
   };
 
-  const result = await dynamoDb.get(params);
-  if ( ! result.Item) {
-    throw new Error("Item not found.");
-  }
-  console.log('Sending back result:', JSON.stringify(result,null,2));
-  // Return the retrieved item
-  return result.Item;
+  await dynamoDb.delete(params);
+
+  return { status: true };
 });
