+++ services/stream-functions/handlers/elasticsearchIndexer.js
